feat: continuação da FundraiserService

+++ colabore/src/main/java/br/com/dbc/devser/colabore/repository/FundraiserRepository.java
@@ -1,9 +1,10 @@
 package br.com.dbc.devser.colabore.repository;
 
 import br.com.dbc.devser.colabore.entity.FundraiserEntity;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FundraiserRepository extends JpaRepository<FundraiserEntity, Integer> {
+public interface FundraiserRepository extends JpaRepository<FundraiserEntity, Long> {
 }
